Refactor search into multiple components
